# CI: reduce puppeteer test time from 34s to 13s