--- a/src/__puppeteer__/.eslintrc.js
+++ b/src/__puppeteer__/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  globals: {
+    jestPuppeteer: 'readonly',
+    page: 'readonly',
+  },
+};