--- a/src/__puppeteer__/routing.test.js
+++ b/src/__puppeteer__/routing.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
-import { configureViewport, goto, sleep } from './utils';
+import { configureViewport, goto } from './utils';
 
-jest.setTimeout(60000);
+jest.setTimeout(30000);
 
 describe('routing', () => {
   beforeAll(async () => {
@@ -15,11 +15,9 @@ describe('routing', () => {
 
   it('load route list', async () => {
     await goto('/a2a0ccea32023010');
-    await sleep(2000);
 
-    // ".DriveList" should be visible
-    const driveList = await page.$('.DriveList');
-    expect(driveList).toBeTruthy();
+    await page.waitForSelector('.DriveList');
+    await page.waitForSelector('.DriveEntry');
 
     // Page should have one ".DriveEntry" element
     const driveEntries = await page.$$('.DriveEntry');
@@ -28,11 +26,16 @@ describe('routing', () => {
 
   it('load route from URL', async () => {
     await goto('/a2a0ccea32023010/1690488081496/1690488851596');
-    await sleep(10000);
 
-    // Should load video with src
-    const video = await page.$('video');
-    expect(video).toBeTruthy();
+    const video = await page.waitForSelector('video');
+
+    // Wait for video src to be set
+    await page.waitForFunction(
+      (vid) => vid.getAttribute('src') !== '',
+      {},
+      video,
+    );
+
     const videoSrc = await page.evaluate((vid) => vid.getAttribute('src'), video);
     expect(videoSrc.startsWith('blob:')).toBeTruthy();
   });