--- a/src/__puppeteer__/routing.test.js
+++ b/src/__puppeteer__/routing.test.js
@@ -38,5 +38,5 @@ describe('routing', () => {
 
     const videoSrc = await page.evaluate((vid) => vid.getAttribute('src'), video);
     expect(videoSrc.startsWith('blob:')).toBeTruthy();
-  });
+  }, 45000);
 });