Replace all `await sleep(...)`s with correct Page.waitForSelector and other query selectors.

Before:
```
 PASS  src/__puppeteer__/offline.test.js (8.013 s)
 PASS  src/__puppeteer__/routing.test.js (22.464 s)
 PASS  src/__puppeteer__/demo.test.js (33.238 s)
```

After:
```
 PASS  src/__puppeteer__/offline.test.js (7.93 s)
 PASS  src/__puppeteer__/demo.test.js (8.148 s)
 PASS  src/__puppeteer__/routing.test.js (12.262 s)
```

Tests should also be less flaky now as the sleeps are not hard-coded for each test.

---

This is much faster to run locally, but GitHub actions still appears to be pretty slow :/